// libs/loans/src/lib/models/modification.model.ts
import { RequestType } from './request-type.model';
export interface Modification extends RequestType {
    hampModification: string;
    fullyCapitalizedAmount: string;
    amortizedAmount: string;
    remainingBalance: string;
    remainingBalanceAction: string;
    interestRate: string;
    termInMonths: string;
    amortizedTerm: string;
    monthlyPIPaymentAmount: string;
    monthlyEscrowPayment: string;
    hoaPayment: string;
    monthlyPitiPayment: string;
    firstPaymentDueDate: string;
    postModificationLTV: string;
    simpleYield: string;
    liquidationType: string;
    liquidationDate: string;
    dtiProposedPITI: string;
    disposableIncome: string;
    grossIncome: string;
    verifiedReserves: string;
};